Fix compilation against older MIT Kerberos and amd64
Clean up some more data type mismatches, mostly just with printf
formats, and fix various problems with portability to old MIT Kerberos.
krb5_kt_free_entry requires AC_CHECK_DECL since older MIT provides
the function.  Flesh out mod-config.h so that mod_webauthldap can use
portable/krb5.h.

+++ modules/webauth/mod_webauth.c
@@ -1533,8 +1533,8 @@ get_session_key(char *token, MWA_REQ_CTXT *rc)
         klen != WA_AES_192 &&
         klen != WA_AES_256) {
         ap_log_error(APLOG_MARK, APLOG_ERR, 0, rc->r->server, 
-                     "mod_webauth: get_session_key: invalid key length: %d",
-                     klen);
+                     "mod_webauth: get_session_key: invalid key length: %lu",
+                     (unsigned long) klen);
         goto cleanup;
     }
